midterm1: Edit for clarity and remove details. (#119)
Idea is that we link to the full recitation notes for each topic. The
midterm notes were getting out of date because people would update the
recitation notes without copying the changes here.

Also the point of this is to jog the students' memory, not to be the
comprehensive source on some topic.