Changes to c-basics.md (#160)
Clarify the sentence preceding the table and make the table show up in a code block instead of a line of text.
Could also format as a table - however, github flavored markdown does not allow for tables without headers, which looks weird in this case. Could also move away from table format, open to suggestions.

Also change the `const` section to refer to `const` as a type qualifier rather than a type modifier (as that is actually what it is). Type modifiers are things like `long`, `short`, `signed`, etc., while type qualifiers are `const`, `volatile`, `restrict`, and `_Atomic`.
Additionally, clean up the final sentence a little.